Converted OperatorSubscribeUntil -> takeUntil
There was one key problem with OperatorSubscribeUntil, which is that
it had the potential to be incompatible with other Operators. This
required you to be careful about when you call it, which isn't great
for a library.

Switching to takeUntil means that onCompleted will be called when
the sequence completes due to a lifecycle event. But that's alright;
in most cases it won't change anything. Also, if someone *does*
need to terminate without onCompleted, they can manually handle
the Subscription and unsubscribe themselves.