Remove generic type declaration for Completable bindings.
Remove the unecessary generic type <T> from the kotlin extensions for
binding to Completable chains. The type declared in the signature is not
being used in the function, and doesn't make sense on a Completable, so
this looks like a copy/paste error.
This fixes the following error, when not specifying a type:

    Type inference failed: Not enough information to infer parameter T in fun <T> Completable.bindUntilEvent(owner: LifecycleOwner, event: Lifecycle.Event): Completable